requirements-dev.txt
flake8
pre-commit
recommonmark
setuptools
sphinx
sphinx-click
sphinxcontrib-napoleon
twine
wheel
